Purpose a pin assignment
This is just a starting point to suggest a pin assignment for this
connector. It would be great if every project with a SOICbite connector
could be used similar, with the same programmer. Anyhow, a "version 1"
probably should be discussed somehow with all the pros and cons. This
one uses the SPI pinout of an AVR AT-Tiny and tries to assign similar
functions from other interfaces to the same pins.
UART TX2 and RX2 could be used as a 2nd (debug) interface in parallel to
e.g. SPI or SWIM.