kicad: add courtyard around symbol
It is good to have courtyard defined as DRC might use it to catch errors
when components are missing. While it doesn't sound that useful for this
particular part, the DRC might actually complain about it being missing
if its turned on for the project.

The courtyard layer is defined on:
https://kicad-pcb.org/libraries/klc/F5.3/

This patch adds the front and back courtyard with the same contour as on
the fab layer.